Add a "Build and Run" feature
The "Build" step calls Unity's BuildPlayer to create an APK in a temp
directory. The "Run" feature pushes that APK to an ADB connected
device and runs it as an instant app.

With Unity 2017.x an AndroidManifest.xml is created/updated in
Assets/Plugins/Android/. With Unity 2018.1+ the manifest is updated
dynamically during the gradle build.