Switch PostBuildCommandLineDialog to Update()
In Unity 5.6 there is no OnGUI() callback after the AppDomain is
reset unless the user interacts with the window, e.g. by clicking
on it. This is the case on both Mac and Windows.

Switch the AppDomain reset check to Update() which is called more
often and works for Unity 5.6.