Add gradle task to build SphereBlast as an instant app
Found a bug in BaseGame.cs that would've been caught with an automated build.
Therefore adding a gradle task that will build the sample as an instant app.

Also noticed that compilation failures weren't failing the gradle build
task, so now throwing an exception whenever there is an APK/AAB build failure.