Fix ApkSigner usage with .NET 4.5
Signing App Bundles with .NET 4.5 was failing due to inclusion of
a byte order marker (BOM) prefix to the keystore password. Change the
output encoding to explicitly not include a BOM.

Remove line indicating when processes are started; it was there
originally for debugging an issue that no longer occurs.

Fix logging of a message when App Bundle building is complete.